requirements.txt
mysql-connector-python== 9.0.0
langchain==0.3.0
langchain-openai==0.2.0
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.9.1
langchain-experimental==0.3.1
openai==1.47.1